requirements.txt
spotipy
json
pandas
pyarrow
sqlalchemy
psycopg2-binary
streamlit
python-dotenv